pkg/ns: enable non-linux builds
